test: Split db/test_storage.py by class
To help with organization.
